更改了一个错误：因为在计算cost时，给出的权重为double,所以 ```cpp struct MGraph{ double *edges[MAX_VALUE];//cost,此处应为double,而不应该是int int iVertexCount,iEdgeCount; };

+++ cdn/floyd_algorithm.cpp
@@ -18,7 +18,7 @@ void Floyd(double **LinkUnitPrice,int &NodeNum_Network, int &LinkNum) {
 		}
 	MGraph mGraph;
 	for (int i = 0; i < MAX_VALUE; i++) {
-		mGraph.edges[i] = new int[MAX_VALUE];
+		mGraph.edges[i] = new double[MAX_VALUE];
 	}
 	mGraph.iVertexCount = NodeNum_Network;
 	mGraph.iEdgeCount = LinkNum;
